// sdk/nodejs/sdk/src/zrok/api/model/addOrganizationMemberRequest.ts
/**
 * zrok
 * zrok client access
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AddOrganizationMemberRequest {
    'organizationToken'?: string;
    'email'?: string;
    'admin'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "organizationToken",
            "baseName": "organizationToken",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "admin",
            "baseName": "admin",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return AddOrganizationMemberRequest.attributeTypeMap;
    }
}