// sdk/nodejs/sdk/src/zrok/api/model/createFrontend201Response.ts
/**
 * zrok
 * zrok client access
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateFrontend201Response {
    'frontendToken'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "frontendToken",
            "baseName": "frontendToken",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateFrontend201Response.attributeTypeMap;
    }
}