// sdk/nodejs/sdk/src/zrok/api/model/createIdentity201Response.ts
/**
 * zrok
 * zrok client access
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateIdentity201Response {
    'identity'?: string;
    'cfg'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "identity",
            "baseName": "identity",
            "type": "string"
        },
        {
            "name": "cfg",
            "baseName": "cfg",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateIdentity201Response.attributeTypeMap;
    }
}